Create final.ino
Scaled down information to match the working of  the three surviving sensors and IMU to reduce noise , combine and analyze the data and then checks against the switch case values for the flex sensors.
The IMU works when the flex sensors are in relaxed mode having no tension on them. The IMU data is directly checked against the values and results are displayed in form of pre-defined text messages.